Load required libraries dynamically
This avoids problems with (probably non-readline) programs that
conflict with them.
For example, the Debian man command seems to setup a seccomp filter,
which caused its processes to fail with SIGSYS and SYS_SECCOMP
when running with ai_readline.so pre-loaded.
This was happening when the ai-readline CURL GNUTLS initialization
called getrandom(2).