Ability to override prompt config section with program specific vals
It is now possible to have system and context in a program section.
For example:

[prompt-sqlite3]
system=You are an assistant who provides SQL commands for sqlite3. ...
context=1

In addition, add verbose debugging errors and more importantly,
if the config does not parse, output the tokens that are responsible
for this problem.

It also adds a verbose message when the prompt is send to OpenAI.
Other backends might need to be updated.